Add `view debug permission` to social_core
In some cases we want to expose a bit more information to developers so
that they can quickly diagnose issues. Information that may not be
understandable to site managers (such as raw field names).

To provide a unified solution for this that does not rely on the
developer user having uid 1 we introduce a new permission. This is
placed in social_core so that we can use it in all modules building on
top of this.